PLAT-96 Moved EmfAttributeColumnHandler
